commit-message-agent/pyproject.toml
[project]
name = "commit-message-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "mcp-use",
  "langchain_openai",
  "portkey_ai",
  "typing",
  "python-dotenv"
]
